ProjectSection returns error when attempted to build as static page, moved to components
